Bug Fixes and Finished Test Cases

+++ Team_A/src/RemoveExpense.java
@@ -38,20 +38,27 @@ public class RemoveExpense extends HttpServlet {
         PrintWriter out = response.getWriter();
 		String html = setupPage();
 		html += removeExpense(user, name);
-		html += "<a href=\\Team_A\\Home_Page.html>Home</a> <br>";
-		html += "</body></html>"; // end of body, html
-		out.println(html);
+		if(html.equals(setupPage())){
+			response.sendRedirect("/Team_A/Create_budget.html");
+		}
+		else {
+			html += "<a href=\\Team_A\\Home_Page.html>Home</a> <br>";
+			html += "</body></html>"; // end of body, html
+			out.println(html);
+		}
 	}
 
 	
 	private String removeExpense(User user, String name) {
 		Budget budget = UtilDB.getBudget(user);
 		String html = "";
-		Expense expense = UtilDB.removeExpense(budget, name);
-		
-		html += String.format("<h2>Expense added to: Budget %s </h2>", budget.getName() );
-		if(expense != null) {
-			html += String.format("%s , %.2f", expense.getName(), expense.getAmount());
+		if(budget != null) {
+			Expense expense = UtilDB.removeExpense(budget, name);
+			
+			html += String.format("<h2>Expense added to: Budget %s </h2>", budget.getName() );
+			if(expense != null) {
+				html += String.format("<p>%s , %.2f</p>\n", expense.getName(), expense.getAmount());
+			}
 		}
 		return html;
 	}
